+++ NEWS.md
@@ -8,6 +8,7 @@
 - add `key.position` option to `timePlot` to control the location of the key.
 - fix labels in `timeVariation` when data for some types is missing.
 - fix `trendLevel` issue due to `dplyr`
+- use `lubridate` package in `timeAverage`
 
 
 # Changes in openair 1.8-6